Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 1680-concatenation-of-consecutive-binary-numbers/1680-concatenation-of-consecutive-binary-numbers.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
 long long M=1e9+7;
     int concatenatedBinary(int n) {
-        long result=0;
+        long long result=0;
         
         for(int i=1;i<=n;i++){
             int digits=log2(i)+1;
